+++ payment/src/main/java/net/optile/payment/ui/widget/input/GroupingTextWatcher.java
@@ -4,27 +4,20 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import android.text.Editable;
-import android.text.TextWatcher;
 import android.widget.EditText;
 
 /**
  * Class for separating 4 digits with a space
  */
-public class GroupingTextWatcher implements TextWatcher {
+public class GroupingTextWatcher extends EditTextWatcher {
 
     private static final char DIVIDER = ' ';
     private static final String STRING_DIVIDER = " ";
 
-    private final EditText editText;
     private final int groupSize;
     private final int dividerModulo;
     private final String inputRegex;
     private final Pattern inputPattern;
-
-    private String previousText = "";
-    private int deleteLength;
-    private int insertLength;
-    private int start;
 
     /**
      * Construct a new GroupingTextWatcher for the given editText
@@ -33,13 +26,11 @@ public class GroupingTextWatcher implements TextWatcher {
      * @param editText for which this class is manipulating the input
      */
     public GroupingTextWatcher(int groupSize, EditText editText) {
+        super(editText);
+
         if (groupSize <= 0) {
             throw new IllegalArgumentException("groupSize must be equal or greater than 1");
         }
-        if (editText == null) {
-            throw new IllegalArgumentException("editText must not be null");
-        }
-        this.editText = editText;
         this.groupSize = groupSize;
         this.dividerModulo = groupSize + 1;
         this.inputRegex = "(\\w{" + groupSize + "}" + DIVIDER + ")*\\w{1," + groupSize + "}";
@@ -50,28 +41,10 @@ public class GroupingTextWatcher implements TextWatcher {
      * {@inheritDoc}
      */
     @Override
-    public void beforeTextChanged(final CharSequence s, final int start, final int count, final int after) {
-        this.previousText = s.toString();
-        this.deleteLength = count;
-        this.insertLength = after;
-        this.start = start;
-    }
+    public void validate(final Editable s) {
+        String afterText = s.toString();
 
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void onTextChanged(final CharSequence s, final int start, final int before, final int count) {
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void afterTextChanged(final Editable s) {
-        String originalString = s.toString();
-
-        if (!previousText.equals(originalString) && !isInputCorrect(originalString)) {
+        if (!previousText.equals(afterText) && !isInputCorrect(afterText)) {
             String newString = previousText.substring(0, start);
             int cursor = start;
 
@@ -81,16 +54,14 @@ public class GroupingTextWatcher implements TextWatcher {
             }
 
             if (insertLength > 0) {
-                newString += originalString.substring(start, start + insertLength);
+                newString += afterText.substring(start, start + insertLength);
                 newString = buildCorrectInput(newString);
                 cursor = newString.length();
             }
             newString += previousText.substring(start + deleteLength);
             s.replace(0, s.length(), buildCorrectInput(newString));
 
-            if (cursor >= 0 && cursor < editText.length()) {
-                editText.setSelection(cursor);
-            }
+            setCursor(cursor);
         }
     }
 
@@ -102,7 +73,7 @@ public class GroupingTextWatcher implements TextWatcher {
      * @param input String to be evaluated
      * @return true if string input is written correctly
      */
-    private boolean isInputCorrect(String input) {
+    private boolean isInputCorrect(final String input) {
         Matcher matcherDot = inputPattern.matcher(input);
         return matcherDot.matches();
     }
@@ -115,7 +86,7 @@ public class GroupingTextWatcher implements TextWatcher {
      * @param input String to be corrected.
      * @return String corrected.
      */
-    private String buildCorrectInput(String input) {
+    private String buildCorrectInput(final String input) {
         StringBuilder sbs = new StringBuilder(
             input.replaceAll(STRING_DIVIDER, ""));